agents/enterprise/multiple-customer-support/ticket-classification-rag/backend/requirements.txt
boto3==1.35.0
botocore==1.35.0
pydantic==2.10.6
pydantic-settings==2.7.0
opensearch-py==2.7.1
requests-aws4auth==1.3.1
tiktoken==0.8.0
PyYAML==6.0.2
